Change decorator for API definitions
